+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/ajax/dto/CreateSampleResponse.java
@@ -1,7 +1,0 @@
-package ca.corefacility.bioinformatics.irida.ria.web.ajax.dto;
-
-/**
- * Abstract class for the response for creating a new sample if a project.
- */
-public abstract class CreateSampleResponse {
-}