+++ src/test/java/ca/corefacility/bioinformatics/irida/ria/unit/web/analysis/AnalysisAjaxControllerTest.java
@@ -35,9 +35,7 @@ import ca.corefacility.bioinformatics.irida.ria.web.analysis.auditing.AnalysisAu
 import ca.corefacility.bioinformatics.irida.ria.web.components.AnalysisOutputFileDownloadManager;
 import ca.corefacility.bioinformatics.irida.ria.web.services.AnalysesListingService;
 import ca.corefacility.bioinformatics.irida.security.permissions.analysis.UpdateAnalysisSubmissionPermission;
-import ca.corefacility.bioinformatics.irida.service.AnalysisSubmissionService;
-import ca.corefacility.bioinformatics.irida.service.ProjectService;
-import ca.corefacility.bioinformatics.irida.service.SequencingObjectService;
+import ca.corefacility.bioinformatics.irida.service.*;
 import ca.corefacility.bioinformatics.irida.service.sample.MetadataTemplateService;
 import ca.corefacility.bioinformatics.irida.service.sample.SampleService;
 import ca.corefacility.bioinformatics.irida.service.user.UserService;
@@ -72,6 +70,9 @@ public class AnalysisAjaxControllerTest {
 	private ExecutionManagerConfig configFileMock;
 	private AnalysisAudit analysisAuditMock;
 	private HttpServletResponse httpServletResponseMock;
+	private AnalysisTypesService analysisTypesServiceMock;
+	private EmailController emailControllerMock;
+
 	/**
 	 * Analysis Output File key names from {@link TestDataFactory#constructAnalysis()}
 	 */
@@ -93,11 +94,13 @@ public class AnalysisAjaxControllerTest {
 		MessageSource messageSourceMock = mock(MessageSource.class);
 		analysisAuditMock = mock(AnalysisAudit.class);
 		httpServletResponseMock = mock(HttpServletResponse.class);
+		analysisTypesServiceMock = mock(AnalysisTypesService.class);
+		emailControllerMock = mock(EmailController.class);
 
 		analysisAjaxController = new AnalysisAjaxController(analysisSubmissionServiceMock, iridaWorkflowsServiceMock,
 				userServiceMock, sampleService, projectServiceMock, updatePermission, metadataTemplateService,
 				sequencingObjectService, analysisSubmissionSampleProcessor,
-				analysisOutputFileDownloadManager, messageSourceMock, configFileMock, analysisAuditMock);
+				analysisOutputFileDownloadManager, messageSourceMock, configFileMock, analysisAuditMock, analysisTypesServiceMock, emailControllerMock);
 
 	}