+++ src/test/java/ca/corefacility/bioinformatics/irida/ria/unit/web/analysis/AnalysisControllerTest.java
@@ -50,8 +50,6 @@ public class AnalysisControllerTest {
 	private AnalysisSubmissionService analysisSubmissionServiceMock;
 	private IridaWorkflowsService iridaWorkflowsServiceMock;
 	private UserService userServiceMock;
-	private EmailController emailControllerMock;
-	private AnalysisAudit analysisAuditMock;
 	private AnalysisTypesService analysisTypesService;
 
 	/**
@@ -65,12 +63,10 @@ public class AnalysisControllerTest {
 		analysisSubmissionServiceMock = mock(AnalysisSubmissionService.class);
 		iridaWorkflowsServiceMock = mock(IridaWorkflowsService.class);
 		userServiceMock = mock(UserService.class);
-		emailControllerMock = mock(EmailController.class);
-		analysisAuditMock = mock(AnalysisAudit.class);
 		analysisTypesService = mock(AnalysisTypesService.class);
 
 		analysisController = new AnalysisController(analysisSubmissionServiceMock, iridaWorkflowsServiceMock,
-				userServiceMock, emailControllerMock, analysisAuditMock, analysisTypesService);
+				userServiceMock);
 
 	}
 
@@ -92,15 +88,13 @@ public class AnalysisControllerTest {
 		when(analysisTypesService.getViewerForAnalysisType(BuiltInAnalysisTypes.PHYLOGENOMICS)).thenReturn(
 				Optional.of("tree"));
 
-		Principal principal = () -> USER_NAME;
-
-		String analysisPage = analysisController.getDetailsPage(submissionId, model, principal);
+		String analysisPage = analysisController.getDetailsPage(submissionId, model);
 		assertEquals("should be analysis page", AnalysisController.ANALYSIS_PAGE, analysisPage);
 
 		assertEquals("Phylogenetic Tree tab should be available", BuiltInAnalysisTypes.PHYLOGENOMICS,
 				model.get("analysisType"));
 
-		assertEquals("submission should be in model", submission, model.get("analysisSubmission"));
+		assertEquals("submission name should be in model", submission.getName(), model.get("analysisName"));
 
 		assertEquals("analysisType should be PHYLOGENOMICS", BuiltInAnalysisTypes.PHYLOGENOMICS,
 				model.get("analysisType"));
@@ -123,15 +117,14 @@ public class AnalysisControllerTest {
 		when(iridaWorkflowsServiceMock.getIridaWorkflowOrUnknown(submission)).thenReturn(iridaWorkflow);
 		when(analysisTypesService.getViewerForAnalysisType(BuiltInAnalysisTypes.PHYLOGENOMICS)).thenReturn(
 				Optional.of("tree"));
-		Principal principal = () -> USER_NAME;
 
-		String analysisPage = analysisController.getDetailsPage(submissionId, model, principal);
+		String analysisPage = analysisController.getDetailsPage(submissionId, model);
 		assertEquals("should be analysis page", AnalysisController.ANALYSIS_PAGE, analysisPage);
 
-		assertFalse("Phylogenetic Tree tab should not be available",
+		assertFalse("Analysis should not be completed",
 				submission.getAnalysisState() == AnalysisState.COMPLETED);
 
-		assertEquals("submission should be in model", submission, model.get("analysisSubmission"));
+		assertEquals("submission name should be in model", submission.getName(), model.get("analysisName"));
 	}
 
 	@Test
@@ -139,8 +132,6 @@ public class AnalysisControllerTest {
 		Long submissionId = 1L;
 		ExtendedModelMap model = new ExtendedModelMap();
 		UUID workflowId = UUID.randomUUID();
-
-		Principal principal = () -> USER_NAME;
 
 		AnalysisSubmission submission = TestDataFactory.constructAnalysisSubmission(workflowId);
 		submission.setAnalysisState(AnalysisState.COMPLETED);
@@ -150,10 +141,10 @@ public class AnalysisControllerTest {
 				createUnknownWorkflow(workflowId));
 		when(analysisTypesService.getViewerForAnalysisType(BuiltInAnalysisTypes.UNKNOWN)).thenReturn(Optional.empty());
 
-		String analysisPage = analysisController.getDetailsPage(submissionId, model, principal);
+		String analysisPage = analysisController.getDetailsPage(submissionId, model);
 		assertEquals("should be analysis page", AnalysisController.ANALYSIS_PAGE, analysisPage);
 
-		assertEquals("submission should be in model", submission, model.get("analysisSubmission"));
+		assertEquals("submission name should be in model", submission.getName(), model.get("analysisName"));
 
 		assertEquals("analysisType should be UNKNOWN", BuiltInAnalysisTypes.UNKNOWN, model.get("analysisType"));
 	}